door + fixed movement
Fixed camera when you switch the door and fixed moving

+++ src/Player.cpp
@@ -1,6 +1,6 @@
 #include "Player.hpp"
 #include <iostream>
-Player::Player(int x, int y, int hp, float speed, float damage, int sizex, int sizey){
+Player::Player(float x, float y, int hp, float speed, float damage, int sizex, int sizey){
     player_shape.setFillColor(sf::Color::Red);
     player_shape.setSize(sf::Vector2f(sizex,sizey));
     player_shape.setPosition(x-(sizex/2), y-(sizey/2));
@@ -17,7 +17,7 @@ void Player::draw(sf::RenderWindow &w){
     
     w.draw(player_shape);
 }
-void Player::move(int xx, int yy){
+void Player::move(float xx, float yy){
     this->player_shape.move(xx, yy);
     x += xx;
     y += yy;
